+++ tests/lab.c
@@ -218,7 +218,7 @@ static colrcv_test_result_t test_colrcv_lab_to_rgb(void) {
     bool success = true;
 
     // convert each colour and compare with output
-    for(size_t i = 0; i < 4; i++) {
+    for(uint8_t i = 0; i < 4; i++) {
         colrcv_rgb_t result;
         colrcv_lab_to_rgb(colours[i].input, &result);
         bool conversion_ok = (
@@ -283,7 +283,7 @@ static colrcv_test_result_t test_colrcv_lab_to_xyz(void) {
     bool success = true;
 
     // convert each colour and compare with output
-    for(size_t i = 0; i < 4; i++) {
+    for(uint8_t i = 0; i < 4; i++) {
         colrcv_xyz_t result;
         colrcv_lab_to_xyz(colours[i].input, &result);
         bool conversion_ok = (
